feat(ride): add `DistanceCalculator`
